[networks] Refactor HTTP encoding
